pretty print for if/else and loops
"The vigor of civilized societies is preserved by the widespread sense that high aims are worth-while. Vigorous societies harbor a certain extravagance of objectives, so that men wander beyond the safe provision of personal gratifications. All strong interests easily become impersonal, the love of a good job well done. There is a sense of harmony about such an accomplishment, the Peace brought by something worth-while."
Alfred North Whitehead, 1963, in The History of Manned Space Flight